// src/app/core/models/Lotery/client.ts
export class Client { 
    id?: number;
    tipoDocumentoIdentidad: string;
    nroDocumentoIdentidad: number;
    email: string;
    primerNombre: string;
    segundoNombre: string;
    primerApellido: string;
    segundoApellido: string;
    estado: string;
    ciudad: string;
    direccion: string;
    codTelefono: string;
    nroTelefono: number;
    }